+++ ec2cmd.go
@@ -8,11 +8,57 @@ import (
 	"github.com/aws/aws-sdk-go/aws"
 	"github.com/aws/aws-sdk-go/aws/session"
 	"github.com/aws/aws-sdk-go/service/ec2"
-
 	"github.com/codegangsta/cli"
 
-	"github.com/reiki4040/peco"
+	"github.com/reiki4040/cstore"
+	myec2 "github.com/reiki4040/rnzoo/ec2"
 )
+
+const (
+	EC2LIST_DESC = `
+     you can set default region by AWS_REGION environment variable.
+
+         export AWS_REGION=ap-northeast-1
+
+     this command make cache file that ec2 info. (default ~/.rnzoo/instance.cache.REGION)
+     second time, you can get ec2 info without access to AWS.
+
+     if you updated ec2(create new instance, stop, start and etc...), need to update cache with -f/--force option.
+
+         ec2list -r ap-northeast-1 -f`
+
+	EC2LIST_USAGE = `show your ec2 infomations with LTSV format.
+
+     show your ec2 info at ap-northeast-1
+
+       rnzoo ec2list -r ap-northeast-1
+
+       instance_id:i-11111111	name:Name tags Value1	state:stopped	public_ip:X.X.X.X	private_ip:Y.Y.Y.Y	instance_type:t2.micro
+       instance_id:i-22222222	name:Name tags Value2	state:running	public_ip:X.X.X.x	private_ip:Y.Y.Y.y	instance_type:m3.large
+       ...
+`
+
+	EC2LIST_FORCE_USAGE  = `reload ec2 (force connect to AWS)`
+	EC2LIST_REGION_USAGE = `specify AWS region name.`
+)
+
+var commandEc2list = cli.Command{
+	Name:        "ec2list",
+	ShortName:   "ls",
+	Usage:       EC2LIST_USAGE,
+	Description: EC2LIST_DESC,
+	Action:      doEc2list,
+	Flags: []cli.Flag{
+		cli.BoolFlag{
+			Name:  OPT_FORCE + ", f",
+			Usage: EC2LIST_FORCE_USAGE,
+		},
+		cli.StringFlag{
+			Name:  OPT_REGION + ", r",
+			Usage: EC2LIST_REGION_USAGE,
+		},
+	},
+}
 
 var commandEc2start = cli.Command{
 	Name:        "ec2start",
@@ -40,32 +86,36 @@ var commandEc2stop = cli.Command{
 	},
 }
 
-func chooseEC2(region string, reload bool) ([]*string, error) {
+func doEc2list(c *cli.Context) {
+	isReload := c.Bool(OPT_FORCE)
+
+	regionName := c.String(OPT_REGION)
+	if regionName == "" {
+		regionName = os.Getenv(ENV_AWS_REGION)
+	}
+
+	if regionName == "" {
+		log.Fatalf("please set region.")
+	}
+
+	err := CreateRnzooDir()
+	if err != nil {
+		log.Printf("can not create rnzoo dir: %s\n", err.Error())
+	}
+
 	h, err := NewRnzooCStoreManager()
 	if err != nil {
-		return nil, err
+		log.Printf("can not load EC2: %s\n", err.Error())
 	}
 
-	ec2list, err := h.LoadChoosableEC2List(region, reload)
+	ec2list, err := h.LoadChoosableEC2List(regionName, isReload)
 	if err != nil {
-		return nil, err
+		log.Printf("can not load EC2: %s\n", err.Error())
 	}
 
-	choices := ConvertChoosableList(ec2list)
-
-	chosens, err := peco.Choose("EC2", "select instances", "", choices)
-	if err != nil {
-		return nil, err
+	for _, i := range ec2list {
+		fmt.Println(i.Choice())
 	}
-
-	ids := make([]*string, 0, len(chosens))
-	for _, c := range chosens {
-		if ec2, ok := c.(*ChoosableEC2); ok {
-			ids = append(ids, aws.String(ec2.InstanceId))
-		}
-	}
-
-	return ids, nil
 }
 
 func doEc2start(c *cli.Context) {
@@ -84,8 +134,13 @@ func doEc2start(c *cli.Context) {
 	instanceId := c.String(OPT_INSTANCE_ID)
 	var ids []*string
 	if instanceId == "" {
-		var err error
-		ids, err = chooseEC2(region, false)
+
+		h, err := NewRnzooCStoreManager()
+		if err != nil {
+			log.Printf("can not load EC2: %s\n", err.Error())
+		}
+
+		ids, err = h.ChooseEC2(region, false)
 		if err != nil {
 			log.Fatalf("error during selecting: %s", err.Error())
 			return
@@ -132,8 +187,13 @@ func doEc2stop(c *cli.Context) {
 	instanceId := c.String(OPT_INSTANCE_ID)
 	var ids []*string
 	if instanceId == "" {
-		var err error
-		ids, err = chooseEC2(region, false)
+
+		h, err := NewRnzooCStoreManager()
+		if err != nil {
+			log.Printf("can not load EC2: %s\n", err.Error())
+		}
+
+		ids, err = h.ChooseEC2(region, false)
 		if err != nil {
 			log.Fatalf("error during selecting: %s", err.Error())
 			return
@@ -163,6 +223,16 @@ func doEc2stop(c *cli.Context) {
 	}
 
 	log.Printf("finished stopping.")
+}
+
+func NewRnzooCStoreManager() (*myec2.EC2Handler, error) {
+	dirPath := GetRnzooDir()
+	m, err := cstore.NewManager("rnzoo", dirPath)
+	if err != nil {
+		return nil, err
+	}
+
+	return myec2.NewEC2Handler(m), nil
 }
 
 func GetDefaultRegion() (string, error) {
